Move svg taks into separate build script
Upload to a single zip file
