+++ .github/workflows/buildsvgs.yml
@@ -22,10 +22,8 @@ jobs:
       - name: Fetch icon sources
         run: cd packages/react-icons && yarn fetch
       - name: Build icons
-        run: cd packages/react-icons && yarn build
-      - name: create icon tarball
-        run: tar cvzf react-icons-svgs.tar.gz -C packages/_react-icons_all-svgs .
+        run: cd packages/react-icons && yarn build-svgs
       - uses: actions/upload-artifact@v3
         with:
           name: react-icons-svgs
-          path: react-icons-svgs.tar.gz
+          path: packages/_react-icons_all-svgs/
